refactored LLM training to separate script
